+++ processor/src/host/mod.rs
@@ -1,12 +1,8 @@
 use alloc::sync::Arc;
 
-use vm_core::{
-    DebugOptions, Word,
-    mast::{MastForest, MastNodeExt},
-};
+use vm_core::{DebugOptions, Felt, Word, mast::MastForest};
 
-use super::{ExecutionError, ProcessState};
-use crate::{Felt, KvMap, MemAdviceProvider, errors::ErrorContext};
+use crate::{ExecutionError, KvMap, ProcessState, errors::ErrorContext};
 
 pub(super) mod advice;
 use advice::AdviceProvider;
@@ -28,20 +24,18 @@ pub use mast_forest_store::{MastForestStore, MemMastForestStore};
 /// 3. handling advice events (which internally mutates the advice provider), and
 /// 4. handling debug and trace events.
 pub trait Host {
-    type AdviceProvider: AdviceProvider;
-
     // REQUIRED METHODS
     // --------------------------------------------------------------------------------------------
 
     /// Returns a reference to the advice provider.
-    fn advice_provider(&self) -> &Self::AdviceProvider;
+    fn advice_provider(&self) -> &AdviceProvider;
 
     /// Returns a mutable reference to the advice provider.
-    fn advice_provider_mut(&mut self) -> &mut Self::AdviceProvider;
+    fn advice_provider_mut(&mut self) -> &mut AdviceProvider;
 
     /// Returns MAST forest corresponding to the specified digest, or None if the MAST forest for
     /// this digest could not be found in this [Host].
-    fn get_mast_forest(&self, node_digest: &Word) -> Option<Arc<MastForest>>;
+    fn get_mast_forest(&mut self, node_digest: &Word) -> Option<Arc<MastForest>>;
 
     // PROVIDED METHODS
     // --------------------------------------------------------------------------------------------
@@ -51,7 +45,7 @@ pub trait Host {
         &mut self,
         _process: ProcessState,
         _event_id: u32,
-        _err_ctx: &ErrorContext<'_, impl MastNodeExt>,
+        _err_ctx: &impl ErrorContext,
     ) -> Result<(), ExecutionError> {
         #[cfg(feature = "std")]
         std::println!(
@@ -97,18 +91,25 @@ impl<H> Host for &mut H
 where
     H: Host,
 {
-    type AdviceProvider = H::AdviceProvider;
-
-    fn advice_provider(&self) -> &Self::AdviceProvider {
+    fn advice_provider(&self) -> &AdviceProvider {
         H::advice_provider(self)
     }
 
-    fn advice_provider_mut(&mut self) -> &mut Self::AdviceProvider {
+    fn advice_provider_mut(&mut self) -> &mut AdviceProvider {
         H::advice_provider_mut(self)
     }
 
-    fn get_mast_forest(&self, node_digest: &Word) -> Option<Arc<MastForest>> {
+    fn get_mast_forest(&mut self, node_digest: &Word) -> Option<Arc<MastForest>> {
         H::get_mast_forest(self, node_digest)
+    }
+
+    fn on_event(
+        &mut self,
+        process: ProcessState,
+        event_id: u32,
+        err_ctx: &impl ErrorContext,
+    ) -> Result<(), ExecutionError> {
+        H::on_event(self, process, event_id, err_ctx)
     }
 
     fn on_debug(
@@ -117,15 +118,6 @@ where
         options: &DebugOptions,
     ) -> Result<(), ExecutionError> {
         H::on_debug(self, process, options)
-    }
-
-    fn on_event(
-        &mut self,
-        process: ProcessState,
-        event_id: u32,
-        err_ctx: &ErrorContext<'_, impl MastNodeExt>,
-    ) -> Result<(), ExecutionError> {
-        H::on_event(self, process, event_id, err_ctx)
     }
 
     fn on_trace(&mut self, process: ProcessState, trace_id: u32) -> Result<(), ExecutionError> {
@@ -141,31 +133,14 @@ where
 // ================================================================================================
 
 /// A default [Host] implementation that provides the essential functionality required by the VM.
-pub struct DefaultHost<A> {
-    adv_provider: A,
+#[derive(Debug, Clone, Default)]
+pub struct DefaultHost {
+    adv_provider: AdviceProvider,
     store: MemMastForestStore,
 }
 
-impl<A: Clone> Clone for DefaultHost<A> {
-    fn clone(&self) -> Self {
-        Self {
-            adv_provider: self.adv_provider.clone(),
-            store: self.store.clone(),
-        }
-    }
-}
-
-impl Default for DefaultHost<MemAdviceProvider> {
-    fn default() -> Self {
-        Self {
-            adv_provider: MemAdviceProvider::default(),
-            store: MemMastForestStore::default(),
-        }
-    }
-}
-
-impl<A: AdviceProvider> DefaultHost<A> {
-    pub fn new(adv_provider: A) -> Self {
+impl DefaultHost {
+    pub fn new(adv_provider: AdviceProvider) -> Self {
         Self {
             adv_provider,
             store: MemMastForestStore::default(),
@@ -194,32 +169,26 @@ impl<A: AdviceProvider> DefaultHost<A> {
     }
 
     #[cfg(any(test, feature = "testing"))]
-    pub fn advice_provider(&self) -> &A {
+    pub fn advice_provider(&self) -> &AdviceProvider {
         &self.adv_provider
     }
 
     #[cfg(any(test, feature = "testing"))]
-    pub fn advice_provider_mut(&mut self) -> &mut A {
+    pub fn advice_provider_mut(&mut self) -> &mut AdviceProvider {
         &mut self.adv_provider
-    }
-
-    pub fn into_inner(self) -> A {
-        self.adv_provider
     }
 }
 
-impl<A: AdviceProvider> Host for DefaultHost<A> {
-    type AdviceProvider = A;
-
-    fn advice_provider(&self) -> &Self::AdviceProvider {
+impl Host for DefaultHost {
+    fn advice_provider(&self) -> &AdviceProvider {
         &self.adv_provider
     }
 
-    fn advice_provider_mut(&mut self) -> &mut Self::AdviceProvider {
+    fn advice_provider_mut(&mut self) -> &mut AdviceProvider {
         &mut self.adv_provider
     }
 
-    fn get_mast_forest(&self, node_digest: &Word) -> Option<Arc<MastForest>> {
+    fn get_mast_forest(&mut self, node_digest: &Word) -> Option<Arc<MastForest>> {
         self.store.get(node_digest)
     }
 
@@ -227,7 +196,7 @@ impl<A: AdviceProvider> Host for DefaultHost<A> {
         &mut self,
         _process: ProcessState,
         _event_id: u32,
-        _err_ctx: &ErrorContext<'_, impl MastNodeExt>,
+        _err_ctx: &impl ErrorContext,
     ) -> Result<(), ExecutionError> {
         #[cfg(feature = "std")]
         std::println!(